rework store into multiple models
